func_test/requirements.txt
Flask==1.0.2
Flask_HTTPAuth==3.2.4
google-cloud-bigquery==1.9.0
